feat(content): inject content programically
